% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_zone.R
\name{calc_dispersion}
\alias{calc_dispersion}
\title{Calculate the dispersion parameter for a quasi-poisson model}
\usage{
calc_dispersion(y, X, beta, stratum_vector)
}
\arguments{
\item{y}{a vector of outcomes}

\item{X}{a matrix of predictors, typically the crossbasis output}

\item{beta}{a vector of coefficients}

\item{stratum_vector}{a vector describing the stratum}
}
\description{
Converted from STATA from armstrong 2014 with guidance from ChatGPT
}
