% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vcov.R
\name{calc_vcov}
\alias{calc_vcov}
\title{Calculate the variance-covariance matrix of a poisson model}
\usage{
calc_vcov(y, X, beta, stratum_vector)
}
\arguments{
\item{y}{a vector of outcomes}

\item{X}{a matrix of predictors, typically the crossbasis output}

\item{beta}{a vector of coefficients}

\item{stratum_vector}{a vector describing the stratum}
}
\description{
https://link.springer.com/article/10.1007/s11222-005-4069-4
https://statomics.github.io/SGA2019/assets/poissonIRWLS-implemented.html#variance-covariance-matrix-of-the-model-parameters
}
\details{
and extended to multi-nomial case using ChatGPT, and confirmed it works.

here are other refs
https://doi.org/10.1111/j.2517-6161.1996.tb02079.x
https://doi.org/10.1093/biomet/68.2.563
}
