% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{condPois_2stage}
\alias{condPois_2stage}
\title{Run a conditional poisson model for across many geographic units}
\usage{
condPois_2stage(
  exposure_matrix,
  outcome_tbl,
  global_cen = NULL,
  verbose = F,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = NULL
)
}
\arguments{
\item{global_cen}{a global centering point}

\item{verbose}{whether to print geo_units as they are run}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single region}
}
\description{
MixMeta
nope its this: https://www.sciencedirect.com/science/article/pii/S0140673614621140?via\%3Dihub
avgtmax and range and random effect for county
unclear, Gasp does both here https://github.com/gasparrini/RiskExtrapolation/blob/3caa5c46a748d6a789021554f435e53f9960d4c1/24_MetaRegression.R
and it goes counrty/city
https://github.com/gasparrini/MCC-SO2/blob/1a158b31bf83f3bbca3449b553a6570280306299/Rcode/06.nonlin.R#L47
nlmeta <- mixmeta(nlcoefall, nlvcovall, random=~1|country/city, data=cities,
control=list(showiter=T, igls.inititer=10))
basically, did Gasp 2015 (temp_mean, temp_iqr, random effect for county)
and included the hierarchicacy structure for ids within countries
https://github.com/gasparrini/2015_gasparrini_EHP_Rcodedata
https://github.com/gasparrini/2019_sera_StatMed_Rcode/blob/master/03.MultilevelMA.R
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATION
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 1
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
MIXMETA
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 2 - BLUP
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
