% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{condPois_sb}
\alias{condPois_sb}
\title{Run a conditional poisson model with Spatial Bayesian inference}
\usage{
condPois_sb(
  exposure_matrix,
  outcomes_tbl,
  shp_sf,
  stan_opts,
  global_cen = NULL,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = NULL,
  strata_min = 0,
  verbose = 0
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{outcomes_tbl}{a data.table of outcomes, created by \code{make_outcome_table}}

\item{shp_sf}{a shapefile object describing the geo_units}

\item{stan_opts}{a list of parameters to send to stan}

\item{global_cen}{a global centering point}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single region}

\item{strata_min}{minimum number of cases per strata}

\item{verbose}{whether to print geo_units as they are run}
}
\description{
https://mc-stan.org/docs/2_20/functions-reference/multinomial-distribution.html
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
IF the outcomes_tbl has a FACTOR, enter a recursive loop
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATION
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 1
==========================================================================
//////////////////////////////////////////////////////////////////////////
////////////////////////////////////////////////////////////////////////////
============================================================================
SETUP inputs
============================================================================
#' /////////////////////////////////////////////////////////////////////////
////////////////////////////////////////////////////////////////////////////
============================================================================
Run STAN
============================================================================
#' /////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 2 - GET FINAL ESTIMATES

sb geo needs coef and vcov
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
