% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPoisSB.R
\name{condPoisSB}
\alias{condPoisSB}
\title{Run a conditional poisson model with Spatial Bayesian inference}
\usage{
condPoisSB(
  exposure_matrix,
  outcomes_tbl,
  shp,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = 50
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{shp}{a shapefile describing the network}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for the entire run}

\item{outcomes}{a data.table of outcomes, created by \code{make_outcome_table}}
}
\description{
https://mc-stan.org/docs/2_20/functions-reference/multinomial-distribution.html
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE CROSSBASIS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
Prepare STAN inputs
==========================================================================
//////////////////////////////////////////////////////////////////////////
////////////////////////////////////////////////////////////////////////////
============================================================================
Run STAN
============================================================================
#' /////////////////////////////////////////////////////////////////////////
}
