% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{checks for cmdstan
copied from https://github.com/rok-cesnovar/misc/blob/master/democmdstanr/R/zzz.R}
\usage{
.onLoad(...)
}
\arguments{
\item{...}{}
}
\description{
checks for cmdstan
copied from https://github.com/rok-cesnovar/misc/blob/master/democmdstanr/R/zzz.R
}
