% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_cp.R
\name{get_centered_cp}
\alias{get_centered_cp}
\title{Internal function to get centered cp objects}
\usage{
get_centered_cp(argvar, xcoef, xvcov, this_exp, x_b, global_cen, cen)
}
\arguments{
\item{cen}{}
}
\description{
Needed a function for this because we do it twice: once for regional RRs
and once for BLUP
}
