% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{getSmat}
\alias{getSmat}
\title{Gets the matrix of strata, used in \code{condPois_sb}}
\usage{
getSmat(strata_vector)
}
\arguments{
\item{strata_vector}{}
}
\description{
Gets the matrix of strata, used in \code{condPois_sb}
}
