% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{indexByComponent}
\alias{indexByComponent}
\title{Title}
\usage{
indexByComponent(x)
}
\arguments{
\item{x}{}
}
\description{
https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
