% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_counties}
\alias{ma_counties}
\title{Massachusetts Towns and Cities (2020 Census)}
\format{
A shapefile with one row per Massachusetts city or town and the
following variables:
\describe{
\item{FIPSID}{Character}
\item{COUNTY20}{Character. 2020 Census county FIPS code.}
\item{AREA_ACRES}{Numeric.}
\item{OBJECTID}{Numeric.}
\item{SHAPE_AREA}{Numeric.}
\item{SHAPE_LEN}{Numeric.}
\item{geometry}{Numeric.}
}
}
\source{
U.S. Census Bureau, 2020 Decennial Census; TIGER/MAF geographic products.
}
\description{
A dataset containing geographic identifiers, administrative classifications,
land and water area, population counts, housing units, and related attributes
for all counties in Massachusetts. See \href{https://www.mass.gov/info-details/massgis-data-counties#downloads}{link}
}
\details{
This dataset is typically used for environmental exposure assessment and
can be joined to town-level geographic or demographic datasets using
\code{COUNTY20}.
}
\keyword{datasets}
