% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_deaths}
\alias{ma_deaths}
\title{Daily Deaths by Town}
\format{
A data frame with one row per date and town and the following variables:
\describe{
\item{date}{Date. Calendar date of the observation.}
\item{daily_deaths}{Numeric. Daily deaths - SIMULATED!}
\item{age_grp}{Character. Either '0-17', '18-64' or '65+'}
\item{sex}{Character. either 'M' or 'F' in this example}
\item{TOWN20}{Character. Name of the Massachusetts town associated with the temperature observation.}
\item{COUNTY20}{Character. Name of the Massachusetts county associated with TOWN20.}
}
}
\description{
A SIMULATED dataset containing \emph{fictional} daily deaths linked to
Massachusetts towns. Each row represents a single day–town observation.
}
\details{
This dataset is typically used for environmental exposure assessment and
can be joined to town-level geographic or demographic datasets using
\code{TOWN20}.
}
\keyword{datasets}
