% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cb_list.R
\name{make_cb_list}
\alias{make_cb_list}
\title{make_cb_list}
\usage{
make_cb_list(x, var_list, argvar_list, lag_list, arglag_list)
}
\arguments{
\item{x}{the data.frame with the var_list data in it}

\item{arglag_list}{}
}
\description{
This function is work in progress, will be updated soon
}
\examples{
x:
var_list:    character vector of columns of X that you use, e.g., c('tmean')
argvar_list: list(set1 = list(fun="bs", degree = 2, knots = list(vk)))
lag_list:    c(5, 30)
arglag_list: list(set1 = list(knots = list(lk))), same fcn as argvar_list
}
