% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgrid.R
\name{make_xgrid}
\alias{make_xgrid}
\title{An internal function to make the xgrid}
\usage{
make_xgrid(data, column_mapping, months_subset = 1:12, dt_by = "day")
}
\arguments{
\item{dt_by}{either by day or by week}
}
\description{
An internal function to make the xgrid
}
