% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{nb2subgraph}
\alias{nb2subgraph}
\title{Title}
\usage{
nb2subgraph(x, c_id, comp_ids, offsets)
}
\arguments{
\item{offsets}{}
}
\description{
https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
