% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_zone.R
\name{single_zone}
\alias{single_zone}
\title{Run a conditional poisson model for a single zone}
\usage{
single_zone(
  exposure_matrix,
  outcomes,
  argvar = NULL,
  arglag = NULL,
  maxlag = 5
)
}
\arguments{
\item{maxlag}{}
}
\description{
Run a conditional poisson model for a single zone
}
