% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_zone.R
\name{calc_dispersion}
\alias{calc_dispersion}
\title{Calculate the dispersion parameter for a quasi-poisson model}
\usage{
calc_dispersion(y, X, beta, stratum_vector)
}
\arguments{
\item{stratum_vector}{}
}
\description{
Calculate the dispersion parameter for a quasi-poisson model
}
