% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_single.R
\name{condPois_single}
\alias{condPois_single}
\title{Run a conditional poisson model for a single geographic unit}
\usage{
condPois_single(
  exposure_matrix,
  outcomes_tbl,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = NULL,
  strata_min = 0,
  global_cen = NULL
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{outcomes_tbl}{a data.table of outcomes, created by \code{make_outcome_table}}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single region}

\item{strata_min}{an integer describing the minimum number of cases for a single strata}

\item{global_cen}{global centering point}
}
\description{
Run a conditional poisson model for a single geographic unit
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE CROSSBASIS for this single zone
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
RUN GNM
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT OBJECTS
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
