% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{make_exposure_matrix}
\alias{make_exposure_matrix}
\title{Function to clean and prepare the exposure data matrix}
\usage{
make_exposure_matrix(
  data,
  column_mapping,
  warm_season_months = 5:9,
  maxgap = 5,
  maxlag = 5,
  grp_level = FALSE
)
}
\arguments{
\item{data}{a dataset of exposures}

\item{column_mapping}{a named list that indicates relevant columns in \code{data}. for the exposure
data table, these need to be one of: c('date', "exposure", 'geo_unit', 'geo_unit_grp')}

\item{warm_season_months}{the warm season months for this region, default is Northern Hemisphere's
May through September (5 through 9)}

\item{maxlag}{the number of lags for the exposure variable (default is 5)}

\item{grp_level}{whether to summarize to the group level or not (default)}

\item{maxgaps}{the maximum allowable missing exposure data gap, to be passed to zoo::na.approx (default is 5)}
}
\value{
a data.table of class("exposure")
}
\description{
Function to clean and prepare the exposure data matrix
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE XGRID
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
FILL NA VALUES
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
GROUP LEVEL SUMMARY?
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
ADD LAGS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
