% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_outcome_table.R
\name{make_outcome_table}
\alias{make_outcome_table}
\title{Function to create the outcome table}
\usage{
make_outcome_table(data, column_mapping, collapse_to = NULL, grp_level = FALSE)
}
\arguments{
\item{column_mapping}{a named list that indicates relevant columns in \code{data}. for the exposure
data table, these need to be one of: c('date', "outcome",'factor, 'geo_unit', 'geo_unit_grp')}

\item{collapse_to}{which factors to collapse across}

\item{grp_level}{whether to summarize to the group level or not (default)}
}
\description{
Function to create the outcome table
}
\examples{
Collapse outcomes to a given factor level, or overall
This is a good first step and you can deal here with missing data in the record
like adding zeros and what not

you also need a place to define MinN? is it here or not at all, just
a warning since that is what is defined later
*********
}
