% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_run.R
\name{single_run}
\alias{single_run}
\title{Run a conditional poisson model for a set of parameters}
\usage{
single_run(
  exposure_matrix,
  outcomes_tbl,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = 50
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{outcomes}{a data.table of outcomes, created by \code{make_outcome_table}}
}
\description{
Run a conditional poisson model for a set of parameters
}
